Some cleanup on the image copy&paste.
Fixed a bug preventing images being copied from web browsers.
Removed image subscription and publication. The actual subscription is never used; the insert method is called to add a image, and the resulting link is added as a message.
This means we don't need to know what images are available, as we have a direct link to them.

Added requirement for users to be logged on to insert an image.